Solve bug #1
https://github.com/pluyckx/kam/issues/1
Bug report:
Tested in v1.0.0-beta1

Kick plugin is not working. I uncomment the 'files'
variable. The plugin appears to be loaded, it outputs
log data. But the configured kick file is not detected.

It appears that the [kick] section header is commented
out. I suggest to uncomment the header in the default
configuration. The plugin can then be enabled/disabled
by commenting out the 'files' variable

Signed-off-by: Philip Luyckx <philip.luyckx@gmail.com>