Fixes #19: Mouse/Keyboard check reports false positives
It seems `udevadm monitor` is reporting more information than expected.
When `udevadm monitor` creates output, the keyboard/mouse checkers think
there was activity and keeps the machine alive.

This commit fixes this, when `udevadm monitor` reports something, the
checkers reload their config, but they check if new keyboards and/or mice
are found. Only if that is true, kam will keep the machine alive.